backend/users/serializers.py
from rest_framework import serializers
from .models import User


class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["age", "gender", "level", "goal", "training_type"]
